Removed old RoomRoute and RoomCanvas components
